nooriq_loox_connect/requirements.txt
websockets==15.0.1
azure-iot-device==2.14.0